#94-refactor/유저 관련 엔티티 수정

+++ src/main/java/com/example/demo/user/model/entity/EducatorInfo.java
@@ -1,8 +1,10 @@
 package com.example.demo.user.model.entity;
 
+import javax.persistence.Entity;
+import javax.persistence.Id;
 
 public class EducatorInfo {
-    private int id;
+    private long id;
     private Educator educator;
     //data
 }
